refactor(option-group-header): convert unit tests to RTL
